Move a bunch of directories/files around for clarity
