--- bin/data/news.json
@@ -1,17 +1,0 @@
-[
-    {
-        "title": "Summer is here for over 500 National Lottery funded projects",
-        "text": "National Lottery Funding will support a number of fun-filled community projects in England this summer. 508 projects will share in over £19 million to bring local people together throughout the summer holidays and beyond.",
-        "link": "https://www.biglotteryfund.org.uk/global-content/press-releases/england/040717_eng_rca4a_500-national-lottery-funded-projects"
-    },
-    {
-        "title": "£1 million Scottish Land Fund boost for local communities",
-        "text": "Seven local communities in different parts of Scotland will soon be reaping the rewards of community ownership, thanks to a Scottish Land Fund investment of over £1 million announced today.",
-        "link": "https://www.biglotteryfund.org.uk/global-content/press-releases/scotland/280617_sco_slf_fund-boost-for-local-communities"
-    },
-    {
-        "title": "Scottish Land Fund to help community buy historic Bannockburn House",
-        "text": "An historic 17th-century mansion where Bonnie Prince Charlie famously resided for part of the Jacobite rebellion, is one step closer to community ownership following a successful bid to the Scottish Land Fund.",
-        "link": "https://www.biglotteryfund.org.uk/global-content/press-releases/scotland/280617_sco_slf_historic-bannockburn-house"
-    }
-]
