+++ bin/scripts/cloudfront.js
@@ -10,8 +10,9 @@ const argv = require('yargs')
     .alias('h', 'help')
     .argv;
 const prompt = require('prompt');
-const routes = require('./routes/routes');
+const routes = require('../../routes/routes');
 const _ = require('lodash');
+const path = require('path');
 const AWS = require('aws-sdk');
 const fs = require('fs');
 const moment = require('moment');
@@ -243,7 +244,8 @@ getDistributionConfig.then((data) => { // fetching the config worked
     // store the old config before changing it, just in case...
     const clone = _.cloneDeep(data);
     const timestamp = moment().format('YYYY-MM-DD-HH-mm-ss');
-    const confPath = `./bin/cloudfront/${timestamp}.json`;
+    // const confPath = `../cloudfront/${timestamp}.json`;
+    const confPath = path.join(__dirname, `../cloudfront/${timestamp}.json`);
     const confData = JSON.stringify(clone, null, 4);
 
     // write config to file for backup
