+++ bin/scripts/deploy.js
