+++ modules/boilerplate/cache.js
@@ -1,5 +1,5 @@
 "use strict";
-const app = require('../server');
+const app = require('../../server');
 const globals = require('./globals');
 const config = require('config');
 const cacheControl = require('express-cache-controller');
